+++ scripts/vcfVetting.py
@@ -1,6 +1,5 @@
-import vcf
-import re
 import argparse
+from vcfParsing import *
 
 parser = argparse.ArgumentParser(description='output file')
 parser.add_argument("path", type=str, help='input vcf file path')
@@ -18,19 +17,6 @@ args = parser.parse_args()
 '''
 
 vcf_reader = vcf.Reader(filename=args.path, compressed=True)
-header_fields = ['CHROM', 'POS', 'ID', 'REF', 'ALT', 'QUAL', 'FILTER', 'INFO', 'FORMAT']
-header_lines = ["##fileformat=" + vcf_reader.metadata['fileformat'] + "\n"]
-
-'''
- vcf 4.2 specs: special character handling
-'''
-
-def _special_character(f):
-    switcher = {-1: 'A',
-                -2: 'G',
-                -3: 'R'}
-    return switcher[f] if f in switcher.keys() else '.'
-
 
 '''
  Process header lines:
@@ -40,60 +26,20 @@ def _special_character(f):
  * use reference id passed as a parameter, do not use id derived from reference metadata field
  * Do not use any lines other than specified
 '''
-
-for flt in vcf_reader.filters:
-    header_lines.append("##FILTER=<ID=" + vcf_reader.filters[flt].id +
-                        ",Description=\"" + vcf_reader.filters[flt].desc + "\">\n")
-add_gt = True    # Add GT and AD formats if missing (strelka-specific)
-swap_nt = False  # Swap sample data if we have TUMOR preceding NORMAL (mutect2-specific)
-
+add_gt = True
 for f in vcf_reader.formats:
-    num = vcf_reader.formats[f].num
-    if vcf_reader.formats[f].num is None:
-        num = '.'
-    elif isinstance(vcf_reader.formats[f].num, int) and vcf_reader.formats[f].num < 0:
-        num = _special_character(vcf_reader.formats[f].num)
     if vcf_reader.formats[f].id == 'GT':
         add_gt = False
-    header_lines.append("##FORMAT=<ID=" + vcf_reader.formats[f].id +
-                        ",Number=" + str(num) + ",Type=" + vcf_reader.formats[f].type +
-                        ",Description=\"" + vcf_reader.formats[f].desc + "\">\n")
-if add_gt:
-    header_lines.append("##FORMAT=<ID=GT,Number=1,Type=String,Description=\""
-                        "Genotype, constructed from SGT INFO via external modification\">\n")
-    header_lines.append("##FORMAT=<ID=AD,Number=.,Type=Integer,Description=\""
-                        "Allelic depths for the ref and alt alleles in the order listed\">\n")
 
-for i in vcf_reader.infos:
-    num = vcf_reader.infos[i].num
-    if vcf_reader.infos[i].num is None:
-        num = '.'
-    elif isinstance(vcf_reader.infos[i].num, int) and vcf_reader.infos[i].num < 0:
-        num = _special_character(vcf_reader.infos[i].num)
-    header_lines.append("##INFO=<ID=" + vcf_reader.infos[i].id +
-                        ",Number=" + str(num) + ",Type=" + vcf_reader.infos[i].type +
-                        ",Description=\"" + vcf_reader.infos[i].desc + "\">\n")
-
-for c in vcf_reader.contigs:
-    if re.search('_', vcf_reader.contigs[c].id):
-        continue
-    header_lines.append("##contig=<ID=" + vcf_reader.contigs[c].id + ",length=" + str(vcf_reader.contigs[c].length) +
-                        ",assembly=" + args.reference + ">\n")
-
+header_lines = generate_header(vcf_reader, args.reference)
+swap_nt = False  # Swap sample data if we have TUMOR preceding NORMAL (mutect2-specific)
 inputHash = {}
-
-'''
-    Preserve metadata for inputs
-'''
 
 """ mutect2 - specific normal swap fix: """
 if len(vcf_reader.samples) == 2 and vcf_reader.samples[1] == 'NORMAL' and len(inputHash) == 0:
     swap_nt = True
 if 'normal_sample' in vcf_reader.metadata.keys() and vcf_reader.samples[1] in vcf_reader.metadata['normal_sample']:
     swap_nt = True
-
-header_fields.extend(['NORMAL', 'TUMOR'])
-header_lines.append("#" + "\t".join(header_fields) + "\n")
 
 '''
   Definitions for helper subroutines which