+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+## Unrealeased - 2023-04-11
+- Added post-processing script, new tag for varmerge-scripts (1.8)
 ## 2.1.1 - 2023-02-15
 - Updated  the default varmerge-script version which fixes an issue with non-existent fields in vcf
 - Updated RTs accordingly